pyproject.toml
[project]
name = "averaged-image"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.4",
    "pillow>=12.2.0",
]
